temp

+++ chassis/core/src/main/java/com/griddynamics/jagger/storage/rdb/HibernateKeyValueStorage.java
@@ -32,8 +32,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Required;
 import org.springframework.orm.hibernate3.support.HibernateDaoSupport;
 
-import java.util.Collection;
-import java.util.List;
+import java.util.*;
 
 public class HibernateKeyValueStorage extends HibernateDaoSupport implements KeyValueStorage {
 
@@ -41,8 +40,19 @@ public class HibernateKeyValueStorage extends HibernateDaoSupport implements Key
 
     private int hibernateBatchSize;
 
+    private int sessionLimit;
+
     public int getHibernateBatchSize() {
         return hibernateBatchSize;
+    }
+
+    public int getSessionLimit() {
+        return sessionLimit;
+    }
+
+    @Required
+    public void setSessionLimit(int sessionLimit) {
+        this.sessionLimit = sessionLimit;
     }
 
     @Required
@@ -90,8 +100,25 @@ public class HibernateKeyValueStorage extends HibernateDaoSupport implements Key
     }
 
     @Override
-    public void deleteAll(){
-        getHibernateTemplate().bulkUpdate("delete from KeyValue");
+    public void deleteAll(String sessionId) {
+        ArrayList<String> sessions = (ArrayList) getHibernateTemplate().find("Select distinct k.sessionId from KeyValue k ORDER by k.sessionId");
+        if (sessions.size() == 0)
+            return;
+        if (sessionLimit < 0) {
+            log.warn("Session count can't be < 0; was get {}", sessionLimit);
+            return;
+        }
+        if (sessionLimit == 0) {
+            log.warn("Session count limit is equal '0', all temporary data about sessions will be delete");
+            getHibernateTemplate().bulkUpdate("delete from KeyValue");
+            return;
+        }
+        if (sessions.size() > sessionLimit) {
+            List<String> sessionForDelete = new ArrayList<String>(sessions.subList(0, (sessions.size() - 1) - sessionLimit));
+            getHibernateTemplate().bulkUpdate("delete from KeyValue where sessionId in (?)", sessionForDelete.toArray());
+        }
+        getHibernateTemplate().bulkUpdate("delete from KeyValue where sessionId=?", sessionId);
+
     }
 
     @SuppressWarnings("unchecked")
@@ -146,6 +173,7 @@ public class HibernateKeyValueStorage extends HibernateDaoSupport implements Key
         keyvalue.setNamespace(namespace.toString());
         keyvalue.setKey(key);
         keyvalue.setData(SerializationUtils.serialize(value));
+        keyvalue.setSessionId(namespace.getValues().get(0));
         return keyvalue;
     }
 }
